Add in functions from Spaces.R: - doaj.aa.trim returns an intersection of the DOAJ list and the AA list with leading/trailing spaces deleted. - doaj.aa.repl returns an intersection of the DOAJ list and the AA list with all spaces deleted. - missing returns a vector of the intersection of all journals that are not on the doaj.aa.trim list but are on the doaj.aa.repl list. These journals have spacing issues within the string name (typically involving spacing conventions around punctuation that differ between lists).
Changes the Clean.Journal.Name to Trimmed.Journal.Name; the specificity reflecting the fact that both the trimmed and the replaced journal names are "cleaned."

Modifies the oa.journals function to include a second loop, running through the small missing list to pick up any publications that match on that list. In this case, it was only two electrical engineering journals.